[MIG] save_translation_file: Migration to 11.0 * auto generated README * remove code used testing
