+++ save_translation_file/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
-# © 2016-TODAY Akretion
+# Copyright 2016-TODAY Akretion
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import save_translation_file
+from . import models
