+++ save_translation_file/__manifest__.py
@@ -1,0 +1,19 @@
+# Copyright 2016-TODAY Akretion
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    "name": "Save translation file",
+    "summary": "Allows developers to easily generate i18n files",
+    "version": "11.0.1.0.0",
+    "category": "Tools",
+    "website": "https://github.com/OCA/server-tools",
+    "author": "Akretion, Odoo Community Association (OCA)",
+    "license": "AGPL-3",
+    "application": False,
+    "installable": True,
+    "depends": [
+        "base",
+    ],
+    "data": [
+        "views/save_translation_file_view.xml",
+    ],
+}
