--- save_translation_file/__openerp__.py
@@ -1,28 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2016-TODAY Akretion
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-{
-    "name": "Save translation file",
-    "summary": "Allows developpers to easily generate i18n files",
-    "version": "8.0.1.0.0",
-    "category": "Tools",
-    "website": "https://www.akretion.com/",
-    "author": "Akretion, Odoo Community Association (OCA)",
-    "license": "AGPL-3",
-    "application": False,
-    "installable": True,
-    "external_dependencies": {
-        "python": [],
-        "bin": [],
-    },
-    "depends": [
-        "base",
-    ],
-    "data": [
-        "save_translation_file_view.xml",
-    ],
-    "demo": [
-    ],
-    "qweb": [
-    ]
-}
