+++ save_translation_file/models/__init__.py
@@ -1,4 +1,3 @@
 # Copyright 2016-TODAY Akretion
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import test_save_translation_file
+from . import save_translation_file
