+++ save_translation_file/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Sylvain Calador <sylvain.calador@akretion.com>
