+++ save_translation_file/readme/DESCRIPTION.rst
@@ -1,0 +1,7 @@
+This module was written for **developers** to easily generate i18n files
+(.po and .pot) from the list of modules, instead of using the native
+configuration "Export Translation" wizard.
+
+- The i18n subdirectory is created if missing in the module.
+- A ".po" file is generated for each installed languages.
+- If a ".po" file exists it is **overwritten** (use with caution).
