+++ save_translation_file/readme/USAGE.rst
@@ -1,0 +1,3 @@
+To use this module, you need to:
+
+* Go to the view list (or form) of installed modules and click the button 'Save translation file'
