--- save_translation_file/save_translation_file_view.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record id="module_form" model="ir.ui.view">
-            <field name="name">ir.module.module.form</field>
-            <field name="model">ir.module.module</field>
-            <field name="inherit_id" ref="base.module_form"/>
-            <field name="arch" type="xml">
-                <button name="button_install_cancel" position="after">
-                    <button name="button_save_translation"
-                        string="Save translation file"
-                        confirm="Overwrite all i18n files for this module ?"
-                        type="object"
-                        states="installed"
-                        icon="terp-translate"
-                    />
-                </button>
-            </field>
-        </record>
-        <record id="module_tree" model="ir.ui.view">
-            <field name="name">ir.module.module.tree</field>
-            <field name="model">ir.module.module</field>
-            <field name="inherit_id" ref="base.module_tree"/>
-            <field name="arch" type="xml">
-                <field name="state" position="after">
-                    <button name="button_save_translation"
-                        string="Save translation file"
-                        confirm="Overwrite all i18n files for this module ?"
-                        type="object"
-                        states="installed"
-                        icon="terp-translate"
-                    />
-                </field>
-            </field>
-        </record>
-    </data>
-</openerp>
