+++ save_translation_file/tests/test_save_translation_file.py
@@ -1,13 +1,12 @@
-# -*- coding: utf-8 -*-
-# © 2016-TODAY Akretion
+# Copyright 2016-TODAY Akretion
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import os
 
-from openerp.modules import get_module_path
-from openerp.tools.misc import get_iso_codes
-from openerp.tools.translate import load_language
-import openerp.tests.common as common
+from odoo.modules import get_module_path
+from odoo.tools.misc import get_iso_codes
+from odoo.tools.translate import load_language
+import odoo.tests.common as common
 
 
 class TestIrModuleModule(common.TransactionCase):
